Reduced the number of texture sampling to one fourth
Size of the precomputed gradient texture is increased by 16MB but the
number required to generate noise at any given 3D position has decreased
from 8 to 2.

The implementation is based on the use of fract() and floor() to unpack
a gradient from only one byte in shader code. Thus a single texel fetch
can generate 4 gradients.